Split the Guice module from the config
Move the getConfig method to SimpleSubmitRulesConfig. Change
NoUnresolvedCommentsRule accordingly.

Remove the complicated constructor from SimpleSubmitRulesModule,
at this point the parameters might not be available, ie: in tests.
Depending on the module sounds like a bad idea, so stop doing it.

Remove the hardcoded plugin name, which might be changed by the
server administrator.

Change-Id: Ic5a6a4875861279ae237ca6438adf5fd7694a29e