Respect ignoreSelfApproval and add end-to-end test
This commit wires up the ignoreSelfApproval bit with the new rule in
Gerrit core. It thereby removes the respective rule from the plugin
as that was moved to core as well.

This commit adds an end-to-end test that configures behavior on the
API and asserts that changes are or are not submittable. While
implementing the end-to-end test, we removed the
AbstractSimpleSubmitRulesIT and replaced it with a project config
update call.

The new PluginIT contains end-to-end tests for reading and writing
API values and for all rules that can be configured. Smaller fixes
to the logic were made where necessary. This mostly covers the
ability to set labels and comment rules independently of each other.

Change-Id: I8cf73155028f23297193346c18c0bad7da4576fd