Respect ignoreSelfApproval and add end-to-end test
This commit wires up the ignoreSelfApproval bit with the new rule in
Gerrit core. It thereby removes the respective rule from the plugin
as that was moved to core as well.

This commit adds an end-to-end test that configures behavior on the
API and asserts that changes are or are not submittable.

Change-Id: I8cf73155028f23297193346c18c0bad7da4576fd