NoUnresolvedCommentsRule: Export the rule name
With change I3f08f189d we added a ruleName field to the SubmitRecord
entity, this is to ease the conversion of submit records to new submit
requirements and give them a meaningful name.

Exporting a meaningful rule name in this change so that it could be used
by Gerrit core.

Change-Id: Icbbc2244ddd492d73d387bd603141af1beb817ff